@@ -1,38 +1,14 @@
 apply plugin: 'com.android.library'
 
-ext {
-    bintrayRepo = 'Croller'
-    bintrayName = 'croller'
 
-    publishedGroupId = 'com.sdsmdg.harjot'
-    libraryName = 'Croller'
-    artifact = 'croller'
-
-    libraryDescription = 'A circular seekbar for Android, with a control knob! (for the lack of a better word).'
-
-    siteUrl = 'https://github.com/harjot-oberai/Croller'
-    gitUrl = 'https://github.com/harjot-oberai/Croller.git'
-
-    libraryVersion = '1.0.7'
-
-    developerId = 'harjot-oberai'
-    developerName = 'Harjot Singh Oberai'
-    developerEmail = 'harjot.oberai@gmail.com'
-
-    licenseName = 'The Apache Software License, Version 2.0'
-    licenseUrl = 'http://www.apache.org/licenses/LICENSE-2.0.txt'
-    allLicenses = ["Apache-2.0"]
-}
-
-version = '1.0.1'
 
 android {
-    compileSdkVersion 27
+    compileSdkVersion 30
     buildToolsVersion '28.0.3'
 
     defaultConfig {
         minSdkVersion 16
-        targetSdkVersion 27
+        targetSdkVersion 30
         versionCode 1
         versionName "1.0"
     }
@@ -46,8 +22,8 @@ android {
 
 dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
-    testImplementation 'junit:junit:4.12'
-    implementation 'com.android.support:appcompat-v7:27.1.1'
+    testImplementation 'junit:junit:4.13.2'
+    implementation 'androidx.appcompat:appcompat:1.3.0'
 }
 
 //apply from: 'https://raw.githubusercontent.com/nuuneoi/JCenter/master/installv1.gradle'
